+++ packages/types/src/interfaces/system/rpc.ts
@@ -1,8 +1,6 @@
 // Copyright 2017-2022 @polkadot/types authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-// As per frontier
-
 import type { DefinitionsRpc } from '../../types';
 
 export const rpc: DefinitionsRpc = {