+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## master
+
+Changes:
+
+- Support for `state_call` definitions via API config (exposed on `api.runtime.*`)
+- Type generation for runtime call definitions (not exposed by default)
+
+
 ## 8.10.1 Jun 26, 2022
 
 Contributed: